refactor: Update FontAwesomeIcon enum values and HtmlIcon class
This commit refactors the FontAwesomeIcon enum values in the `HtmlIcon` class to include the appropriate Font Awesome classes. It also updates the `HtmlIcon` class to use the `fontawesom` attribute instead of the `icon_type` attribute for generating the HTML icon. Additionally, it modifies the `HtmlIcon` class to include the `GITHUB` icon for GitHub URLs. Finally, it adds a sample usage of the `HtmlIcon` class in the `__main__` block.